+++ service-worker.js
@@ -3,7 +3,7 @@
 "use strict";
 
 var CACHE_PREFIX = "monster-ticket-pwa-";
-var CACHE_NAME = CACHE_PREFIX + "74-enterprise-751-qr-order-core-20260717-1";
+var CACHE_NAME = CACHE_PREFIX + "74-enterprise-760-core-20260718-1";
 var OFFLINE_PAGE = "./offline.html";
 var CORE_ASSETS = [
     "./index.html",
@@ -57,6 +57,7 @@ var CORE_ASSETS = [
     "./js/core/enterprise-core.js",
     "./js/core/enterprise-bridge.js",
     "./js/core/enterprise-guards.js",
+    "./js/core/ticket-status-engine.js",
     "./js/display.js",
     "./js/modules/businessMode.js",
     "./js/modules/cart.js",